+++ editor_patch/main.cpp
@@ -460,7 +460,7 @@ CodeInjection texture_name_buffer_overflow_injection2{
 
 // Custom event support
 constexpr int original_event_count = 89;
-constexpr int new_event_count = 19; // must be 1 higher than actual count
+constexpr int new_event_count = 20; // must be 1 higher than actual count
 constexpr int total_event_count = original_event_count + new_event_count;
 std::unique_ptr<const char*[]> extended_event_names; // array to hold original + additional event names
 
@@ -484,6 +484,7 @@ const char* additional_event_names[new_event_count] = {
     "Goal_Gate",
     "Environment_Gate",
     "Inside_Gate",
+    "Anchor_Marker",
     "_dummy"
 };